[#80] Feat: 대시보드 컬럼 스크롤 구현
